refactor callback_handler
