+++ src/handlers/feedback.py
@@ -10,14 +10,36 @@ from src.utils.categories import (
     ACKNOWLEDGMENT_CAPTION,
     ACKNOWLEDGMENT_IMAGE_PATH,
     CATEGORIES,
+    SUBCATEGORIES
 )
 from src.utils.media_utils import save_state  
+
+from src.utils.media_utils import send_or_edit_media, save_state
+from src.keyboard import get_identity_choice_keyboard
+from aiogram.types import CallbackQuery
+
+from src.utils.helpers import safe_decode, save_state
+
 
 logger = setup_logger(__name__)
 
 
+
 async def send_feedback_prompt(bot, user_id, feedback_type, is_named):
-    info = CATEGORIES.get(feedback_type, CATEGORIES["Другое"])
+    # Определяем, откуда брать инфо и текст
+    if feedback_type in SUBCATEGORIES:
+        info = SUBCATEGORIES[feedback_type]
+        message_text = info.text  # просто текст из подкатегорий
+    elif feedback_type in CATEGORIES:
+        info = CATEGORIES[feedback_type]
+        if feedback_type == "Другое":
+            message_text = info.text
+        else:
+            message_text = f"Опиши проблему по теме '{feedback_type}':\n\n{info.text}"
+    else:
+        info = CATEGORIES["Другое"]
+        message_text = info.text
+
     state = await redis_client.hgetall(f"user_state:{user_id}")
     image_msg_id = int(state.get("image_message_id", 0))
     text_msg_id = int(state.get("menu_message_id", 0))
@@ -36,7 +58,7 @@ async def send_feedback_prompt(bot, user_id, feedback_type, is_named):
             await bot.edit_message_text(
                 chat_id=user_id,
                 message_id=text_msg_id,
-                text=f"Опиши проблему по теме '{feedback_type}':",
+                text=message_text,
                 reply_markup=None
             )
             await save_state(user_id, prompt_message_id=text_msg_id)
@@ -49,20 +71,60 @@ async def send_feedback_prompt(bot, user_id, feedback_type, is_named):
         )
         text_msg = await bot.send_message(
             chat_id=user_id,
-            text=f"Опиши проблему по теме '{feedback_type}':"
+            text=message_text
         )
-        # Сохраняем оба id и prompt_message_id одним вызовом
-        await save_state(user_id,
-                         image_message_id=image_msg.message_id,
-                         menu_message_id=text_msg.message_id,
-                         prompt_message_id=text_msg.message_id)
+        await save_state(
+            user_id,
+            image_message_id=image_msg.message_id,
+            menu_message_id=text_msg.message_id,
+            prompt_message_id=text_msg.message_id
+        )
 
     logger.info(f"Feedback prompt sent to user {user_id} (named={is_named}) for type {feedback_type}")
+
+async def handle_feedback_choice(callback: CallbackQuery, data: str):
+    user_id = callback.from_user.id
+    if not await can_create_new_feedback(user_id):
+        await callback.answer(
+            "❗️ У вас уже есть открытое обращение. Дождитесь ответа перед созданием нового. ❗️",
+            show_alert=True
+        )
+        logger.info(f"User {user_id} attempted to start new feedback while blocked")
+        return
+
+    await redis_client.set(f"feedback_type:{user_id}", data, ex=300)
+
+    msg = await send_or_edit_media(
+        callback,
+        CATEGORIES.get(data, CATEGORIES["Другое"]).image,
+        photo_caption="",  
+        text="Хочешь остаться анонимом или указать своё?",
+        reply_markup=get_identity_choice_keyboard()
+    )
+    await save_state(user_id, menu_message_id=msg.message_id)
+    await callback.answer()
+
+
+async def handle_send_identity_choice(callback: CallbackQuery, data: str):
+    user_id = callback.from_user.id
+    bot = callback.message.bot
+
+    feedback_type = await redis_client.get(f"feedback_type:{user_id}")
+    if not feedback_type:
+        await callback.answer("Что-то пошло не так. Попробуй ещё раз.", show_alert=True)
+        return
+
+    decoded_type = safe_decode(feedback_type)
+    is_named = data == "send_named"
+    await save_state(user_id, type=decoded_type, is_named=is_named)
+
+    await send_feedback_prompt(bot, user_id, decoded_type, is_named)
+    await callback.answer()
 
 
 async def feedback_message_handler(message: Message):
     user_id = message.from_user.id
-    feedback_key = f"user_state:{user_id}"  # обновлённый ключ prefix!
+    feedback_key = f"user_state:{user_id}"
 
     feedback = await redis_client.hgetall(feedback_key)
     if not feedback:
@@ -128,7 +190,7 @@ async def feedback_message_handler(message: Message):
 
     if prompt_message_id:
         try:
-            await message.bot.delete_message(chat_id=message.chat.id, message_id=int(prompt_message_id))
+            await message.bot.delete_message(chat_id=user_id, message_id=int(prompt_message_id))
         except Exception as e:
             logger.warning(f"Failed to delete prompt message: {e}")
 
@@ -144,15 +206,10 @@ async def feedback_message_handler(message: Message):
     )
 
     if menu_message_id:
+        logger.info(f"Trying to delete old menu message: chat_id={user_id}, message_id={menu_message_id}")
         try:
-            await message.bot.edit_message_media(
-                chat_id=message.chat.id,
-                message_id=int(menu_message_id),
-                media=InputMediaPhoto(media=ack_photo, caption=ack_caption),
-                reply_markup=back_btn
-            )
-            return
+            await message.bot.delete_message(chat_id=user_id, message_id=int(menu_message_id))
         except Exception as e:
-            logger.warning(f"Failed to edit menu message: {e}")
+            logger.warning(f"Failed to delete old menu message: {e}")
 
     await message.answer_photo(photo=ack_photo, caption=ack_caption, reply_markup=back_btn)
